@@ -5,66 +5,52 @@
 
 	<groupId>com.gsdd</groupId>
 	<artifactId>pc-stock</artifactId>
-	<version>1.0.001</version>
+	<version>2.0.1</version>
 	<packaging>jar</packaging>
 
 	<name>pc-stock</name>
 	<organization>
 		<name>GSDD</name>
 	</organization>
-	
-	<repositories>
-		<repository>
-			<id>central</id>
-			<url>https://repo.maven.apache.org/maven2</url>
-		</repository>
-		<repository>
-			<id>xuggler.repo</id>
-			<url>https://www.dcm4che.org/maven2/</url>
-		</repository>
-	</repositories>
 
 	<properties>
 		<project.encoding>UTF-8</project.encoding>
 		<project.build.sourceEncoding>${project.encoding}</project.build.sourceEncoding>
 		<project.build.targetEncoding>${project.encoding}</project.build.targetEncoding>
 		<project.reporting.outputEncoding>${project.encoding}</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
-
+		<java.version>17</java.version>
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
-		<mvn.compiler.plugin.version>3.8.1</mvn.compiler.plugin.version>
-		<mvn.jacoco.plugin.version>0.8.7</mvn.jacoco.plugin.version>
-		<mvn.failsafe.plugin.version>3.0.0-M4</mvn.failsafe.plugin.version>
-		<mvn.surefire.plugin.version>3.0.0-M4</mvn.surefire.plugin.version>
-		<mvn.javadoc.plugin.version>3.1.1</mvn.javadoc.plugin.version>
-		<mvn.format.plugin.version>2.10.0</mvn.format.plugin.version>
-		<mvn.dependency.plugin.version>3.1.1</mvn.dependency.plugin.version>
-		<mvn.source.plugin.version>3.2.0</mvn.source.plugin.version>
+		<mvn.compiler.plugin.version>3.10.1</mvn.compiler.plugin.version>
+		<mvn.jacoco.plugin.version>0.8.8</mvn.jacoco.plugin.version>
+		<mvn.failsafe.plugin.version>3.0.0-M5</mvn.failsafe.plugin.version>
+		<mvn.surefire.plugin.version>3.0.0-M5</mvn.surefire.plugin.version>
+		<mvn.javadoc.plugin.version>3.3.2</mvn.javadoc.plugin.version>
+		<mvn.spotless.plugin.version>2.22.1</mvn.spotless.plugin.version>
+		<mvn.checkstyle.plugin.version>3.1.2</mvn.checkstyle.plugin.version>
+		<mvn.dependency-check.plugin.version>7.1.0</mvn.dependency-check.plugin.version>
+		<mvn.dependency.plugin.version>3.3.0</mvn.dependency.plugin.version>
+		<mvn.source.plugin.version>3.2.1</mvn.source.plugin.version>
 
 		<!-- If true skips UT test execution -->
 		<skip.unit.tests>false</skip.unit.tests>
 		<!-- If true skips IT test execution -->
 		<skip.integration.tests>false</skip.integration.tests>
 
-		<!-- JaCoCo props -->
-		<jacoco.ut.exec>${project.build.directory}/coverage-reports/jacoco-ut.exec</jacoco.ut.exec>
-		<jacoco.ut.reportPath>${project.build.directory}/coverage-reports/jacoco-ut</jacoco.ut.reportPath>
-		<jacoco.it.exec>${project.build.directory}/coverage-reports/jacoco-it.exec</jacoco.it.exec>
-		<jacoco.it.reportPath>${project.build.directory}/coverage-reports/jacoco-it</jacoco.it.reportPath>
+		<checkstyle.version>10.1</checkstyle.version>
 
-		<lombok.version>1.18.22</lombok.version>
+		<lombok.version>1.18.24</lombok.version>
 		<!-- Test -->
-		<junit.version>5.7.2</junit.version>
-		<mockito.version>3.9.0</mockito.version>
+		<junit.version>5.9.0</junit.version>
+		<mockito.version>4.7.0</mockito.version>
 		<!-- Logger api -->
-		<slf4j.version>1.7.32</slf4j.version>
+		<slf4j.version>1.7.36</slf4j.version>
 		<!-- Logger impl -->
-		<log4j.version>2.16.0</log4j.version>
-		<commons.io.version>2.6</commons.io.version>
+		<log4j.version>2.18.0</log4j.version>
+		<commons.io.version>2.11.0</commons.io.version>
 		<xuggler.version>5.4</xuggler.version>
 		<!-- Dependencias internas -->
-		<file.util.version>1.0.001</file.util.version>
+		<file.util.version>2.0.1</file.util.version>
 	</properties>
 
 	<dependencies>
