@@ -19,11 +19,14 @@ public final class TableModelConstants {
   public static final String QUANTITY = PCStockLanguage.getMessageByLocale("tbl.quantity.column");
   public static final String SIZE = PCStockLanguage.getMessageByLocale("tbl.size.column");
   public static final String RES = PCStockLanguage.getMessageByLocale("tbl.res.column");
-  protected static final String[] COMPARE_TABLE_TITLES = {TableModelConstants.NAME,
-      TableModelConstants.C_MAIN, TableModelConstants.C_SND, TableModelConstants.C_DIF};
+  protected static final String[] COMPARE_TABLE_TITLES = {
+    TableModelConstants.NAME,
+    TableModelConstants.C_MAIN,
+    TableModelConstants.C_SND,
+    TableModelConstants.C_DIF
+  };
 
   public static String[] getCompareTableTitles() {
     return COMPARE_TABLE_TITLES;
   }
-
 }
