@@ -1,15 +1,16 @@
 package com.gsdd.pcstock.ui.model;
 
-import javax.swing.table.DefaultTableModel;
 import com.gsdd.pcstock.constants.TableModelConstants;
+import javax.swing.table.DefaultTableModel;
 
 public class SeeDetailedModel extends DefaultTableModel {
 
   private static final long serialVersionUID = 2373032138268517497L;
 
   private static final Class<?>[] TYPES = new Class[] {Object.class, Object.class, Object.class};
-  private static final String[] COLUMNS_RES =
-      {TableModelConstants.NAME, TableModelConstants.SIZE, TableModelConstants.RES};
+  private static final String[] COLUMNS_RES = {
+    TableModelConstants.NAME, TableModelConstants.SIZE, TableModelConstants.RES
+  };
   private static final String[] COLUMNS = {TableModelConstants.NAME, TableModelConstants.SIZE};
 
   public SeeDetailedModel(Object[][] data, Object[] columnNames) {
@@ -24,5 +25,4 @@ public class SeeDetailedModel extends DefaultTableModel {
   public static String[] getColumns(boolean withResolution) {
     return withResolution ? COLUMNS_RES : COLUMNS;
   }
-
 }
