@@ -1,14 +1,14 @@
 package com.gsdd.pcstock.util;
 
+import com.gsdd.pcstock.model.DetailedFile;
 import java.io.File;
 import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
-import org.apache.commons.io.FileUtils;
-import com.gsdd.pcstock.model.DetailedFile;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FileUtils;
 
 @Setter
 @Getter
@@ -20,12 +20,20 @@ public class DetailedFileDirectoryHelper extends AbstractDirectoryHelper<Detaile
   @Override
   public List<DetailedFile> traverseDirectories(String[] filters, List<File> directories) {
     List<DetailedFile> fileList = new CopyOnWriteArrayList<>();
-    directories.parallelStream().forEach(
-        directory -> FileUtils.listFiles(directory, filters, true).stream().forEach(file -> {
-          log.debug("fp > {}", file.getAbsolutePath());
-          fileList.add(DetailedFile.builder().name(file.getName()).size(file.length())
-              .resolution(extractResolution(file.getAbsolutePath())).build());
-        }));
+    directories.parallelStream()
+        .forEach(
+            directory ->
+                FileUtils.listFiles(directory, filters, true).stream()
+                    .forEach(
+                        file -> {
+                          log.debug("fp > {}", file.getAbsolutePath());
+                          fileList.add(
+                              DetailedFile.builder()
+                                  .name(file.getName())
+                                  .size(file.length())
+                                  .resolution(extractResolution(file.getAbsolutePath()))
+                                  .build());
+                        }));
     Collections.sort(fileList);
     return fileList;
   }
@@ -46,5 +54,4 @@ public class DetailedFileDirectoryHelper extends AbstractDirectoryHelper<Detaile
   public Long extractAccountField(DetailedFile pcStockFile) {
     return pcStockFile.getSize();
   }
-
 }
